# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Salary Manager",
        "description": "CRUD&Permisson+Role system.",
        "contact": {
            "email": "Hexagon14@mail.ru"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Main API Server"
        }
    ],
    "paths": {
...
        "/api/currency_record": {
            "get": {
                "tags": [
                    "CurrencyRecords"
                ],
                "summary": "Get list of currency records.",
                "description": "Returns list of currency records.",
                "operationId": "getCurrencyRecordList",
                "responses": {
                    "200": {
                        "description": "Returns list of currency records."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "CurrencyRecords"
                ],
                "summary": "Creates currency list.",
                "description": "Creates currency list.",
                "operationId": "createCurrencyRecord",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "company_id": {
                                        "type": "integer"
                                    },
                                    "project_salary": {
                                        "type": "integer"
                                    },
                                    "currency_id": {
                                        "type": "integer"
                                    },
                                    "bank_rate": {
                                        "type": "integer"
                                    },
                                    "tax_rate": {
                                        "type": "integer"
                                    },
                                    "net": {
                                        "type": "integer"
                                    },
                                    "month": {
                                        "type": "integer"
                                    },
                                    "operation_date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "company_id": "4",
                                    "project_salary": "50000",
                                    "currency_id": "1",
                                    "bank_rate": "5",
                                    "tax_rate": "8",
                                    "net": "6",
                                    "month": "3",
                                    "operation_date": "2006-11-22 12:00:00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Returns currency record."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
        "/api/currency_record/{id}": {
            "get": {
                "tags": [
                    "CurrencyRecords"
                ],
                "summary": "Get one currency record.",
                "description": "Returns one currency record.",
                "operationId": "getCurrencyRecord",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currency record id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Returns currency record."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "CurrencyRecords"
                ],
                "summary": "Deletes currency record.",
                "description": "Deletes currency record.",
                "operationId": "deleteCurrencyRecord",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currency record id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Currency record deleted."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "CurrencyRecords"
                ],
                "summary": "Updates currency record.",
                "description": "Updates currency record parameters that were given in request.",
                "operationId": "updateCurrencyRecord",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Currency record id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "company_id": {
                                        "type": "integer"
                                    },
                                    "project_salary": {
                                        "type": "integer"
                                    },
                                    "currency_id": {
                                        "type": "integer"
                                    },
                                    "bank_rate": {
                                        "type": "integer"
                                    },
                                    "tax_rate": {
                                        "type": "integer"
                                    },
                                    "net": {
                                        "type": "integer"
                                    },
                                    "month": {
                                        "type": "integer"
                                    },
                                    "operation_date": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "company_id": "4",
                                    "project_salary": "50000",
                                    "currency_id": "1",
                                    "bank_rate": "5",
                                    "tax_rate": "8",
                                    "net": "6",
                                    "month": "3",
                                    "operation_date": "2006-11-22 12:00:00"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Returns currency record."
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "Bearer Token": []
                    }
                ]
            }
        },
...
    },
...
    "components": {
        "securitySchemes": {
            "Bearer Token": {
                "type": "apiKey",
                "description": "Laravel passport security.",
                "in": "header",
                "name": "Authorization",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://localhost/auth/login",
                        "tokenUrl": "http://localhost/auth/token",
                        "refreshUrl": "http://localhostauth/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}